Added get by id

+++ src/main/java/com/gippies/markable/greeting/GreetingRepository.java
@@ -2,6 +2,6 @@ package com.gippies.markable.greeting;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GreetingRepository extends JpaRepository<Greeting, Integer> {
+public interface GreetingRepository extends JpaRepository<Greeting, Long> {
 
 }
